Added data sideloading
This allows you to return a JSON response like:

  {
    person: {
      id: 1,
      name: "Yehuda",
      phone_numbers: [ 1 ]
    },
    phone_numbers: [
      { id: 1, number: "555-5555" }
    ]
  }

You can sideload data in any response. If the
data has the same key as an association of the
object being loaded or updated, ember-data will
assume the type specified for that association.

If you want to include unrelated data, you can
specify a mapping property on your adapter:

  var adapter = DS.RESTAdapter.create({
    mapping: {
      summary: App.Summary
    }
  });

This will mean that any time ember-data sees an
unexpected `summary` attribute on the response
JSON, it will load the data as hashes for the
`App.Summary` type.