Use DataProxy & toJSON instead of raw data hash
Previously, a record's data was stored in a hash
that was referenced by both the store and the
record. Now that we have split the "canonical"
source of data and the pending changes that have
not yet been committed, we have created an
internal DataProxy object that coalesces these
two hashes into a transparent object that will
return the correct value when asked.

As part of this change, `data` semantically no
longer means "the data hash to be sent to the
server." In this case, you should use the
`toJSON()` method.