Fix already-materialized associations after `load`
This commit ensures that associations that were
already materialized are properly updated if
they were materialized while their parent record
was still loading.

Previously, this was managed by making the
association computed property depend on `data`,
but this would not cause any references to the
association itself to be updated. This commit
uses the fact that associations are `ArrayProxy`s
to update they `content` in place, thus updating
any existing references as well.