Makes model/record terminology consistent.
Previously, there was some confusion internally
about what was a record and what was a model. This
commit makes the terminology consistent. A model
is a schema definition that creates a subclass of
DS.Model. Instances of those subclasses are called
records. A record corresponds to a single entry
in the persistence layer, as loaded by the
adapter.

To support this change, DS.ModelArray and its
subclasses are now named DS.RecordArray. Since
you should not be using these class names directly
(but rather have them returned by store methods
such as find() and filter()), this should not be
a breaking API change and thus we have not updated
the API revision number.