Fixes bug when record is modified then deleted
Because we defer updating record arrays to the
end of the run loop, but we immediately remove
records from record arrays if they are deleted,
there existed a race condition where, if in the
same run loop, a record was modified and then
deleted.

In that case, the record would be removed from the
record array, but at the end of the run loop, it
would be re-evaluated and potentially added back
to filtered record arrays.

This commit updates the test to take the deferred
scheduling into account, and does not bother
updating record arrays with records that have
been deleted.