Add hook to generate client-side IDs
Previously, the store assumed that record IDs
would always be generated on the server, and
did not update bookkeeping information about
server IDs until the `didCreateRecord` method
was called.

Now, adapters may optionally implement the
`generateIdForRecord` method, which is invoked
each time a new record is created. If this method
exists, its return value will be used as the
record's ID, and internal store bookkeeping will
be updated.

This commit also adds some cursory documentation
to the DS.Adapter file.