WIP do more stuff in the adapter
The most important change here is that filters now
materialize records.

This is necessary because the records no longer
store their backing data as a serialized hash.

The long-term perfz plan (for situations with huge
numbers of data hashes) is to integrate with
something like crossfilter to filter out unneeded
records before loading into the store.