Deal with deletion situation
It turned out to be possible for a relationship
change to exist even if no child record existed.

This happened if a parent with unmaterialized
children was deleted. Because of this situation,
we needed to make the relationship change objects
more standalone, and change the infrastructure to
use clientIds internally instead of expecting at
least the child to be materialized.

This change is somewhat destabilizing, and I
realized that he don't have coverage of
didSaveRecord(<no hash>), but it's necessary.

This change also makes several public methods on
OneToManyChange, which is passed to the adapter,
and requires that adapters acknowledge updates to
the relationships, because of the possibility that
a relationship change can need to be persisted
even if the associated record is not loaded (i.e.
the above deletion case).