Fix for relationship acknowledgement issues
This is an interim fix that should work for
the RESTAdapter but doesn't have enough
conveniences in DS.Adapter to be a good API for
general purpose.

We plan to add mixins (DS.ForeignKey, etc.) for
the common use-cases so that adapters typically
don't have to do this kind of bookkeeping.