Fix hasMany association access during initialization
Accessing hasMany associations in a model's initializer was triggering a
second materialization: first when the hasMany access was attempted, then
again when the store finished initializing the new record. The second
materialization erroneously reset all of the model's data.

    App.Post = DS.Model.extend({
      comments: DS.hasMany(App.Comment),

      init: function() {
        this._super.apply(this, arguments);

        var comments = this.get('comments');

        // ...
      }
    })

This error also occurred when setting observers on a hasMany
association, as this would also trigger a hasMany access during
initialization (in order to add the array observer):

    App.Post = DS.Model.extend({
      comments: DS.hasMany(App.Comment),

      commentsDidChange: function() {
        // ...
      }.observes('comments.@each')
    })