Implements adapter dirtiness hooks
This commit adds the ability for the adapter to
specify which records should be marked dirty when
changes occur, such as when an attribute or
relationship changes.

If you are using the REST adapter, it should
not introduce any breaking changes.

For more information about the rationale
behind this change, please see
BREAKING_CHANGES.md, revision 9.