Implement DS.Store.reset()
This method is primarily used in tests. The reset method undos
everything internally and goes back to how it was when it was
initialized. This does not undo anything committed. It also call's
the adapter's reset method. You can use this with the fixture adapter
to reset the store between tests.

DS.Store.init() calls reset(). This also calls the adapter's reset
method. This means the adapter will be created at instantiation time and
not when first referenced (via the `_adapter` cp). This change made the
embeddded loading tests fail. The tests failed because they added maps
to the adapter after the store was initialized. This is a problem
because the maps are added to the serializer when the adapter is
created. I changed the tests to define all the maps upfront on the
adapter then create the store. I think that instantiating the adapter
when the store is initialized is a much safer strategy then having it
happen at random.