Add notDirty mapping option
Adding a notDirty property to a model property via the map API, it
allows the record property value be updated and cause the record does not
become dirty.

```js
    Adapter.map(App.Person, {
      firstName: { notDirty: true }
    });
```

This feature can be used for example, whenever a client will not
update these model properties to the remote server, although it updates
its local copy to improve the application user experience.