The owner record is now an argument passed to findMany on the adapter
While writing an adapter for my django backend I was unable to
build a valid url for my hasMany relationships because the owner
was not easily accessible. Going forward the owner will now be an
argument available for custom adapters like the one I ended up writing.

example of how my REST api would lookup hasMany relationships:

If I have a person with many occupations I would look them up like so

/people/1/occupations/

Notice the id 1 above is actually the owner of the occupations (not a
occupation id from the ids array that is passed into findMany)

Instead of passing each id like the built in RESTAdapter the url above
just expects the owner will know how to find each occupation