Allow to override getting attribute on a DS.Model
Currently there is no easy way to override getAttr function, which is
used in the computed property returned by DS.attr and therefore it's
hard hook into the moment of getting attribute.

This commit moves getAttr into the model to allow access to this
function. For example:

    var Person = DS.Model.extend({
      name: DS.attr('string'),

      getAttr: function(key, options) {
        console.log("Trying to feth attribute '" + key + "'.);
      }
    });

    // assuming that person is of Person type
    person.get('name')

Such code will output following text to the console:

    Trying to fetch attribute 'name'.