Persist errors object on models.
This is closer to what people expect, and allows users to add observers
on the
errors object itself rather than on the model.

The errors object is cleared when the model becomes in flight, which
supports
"base" errors not associated with a property (and therefore not cleared
on
property set).

[Fixes #852].