RESTAdapter Unit Test Refactor - Bad
!!! DO NOT MERGE THIS BRANCH !!!

This branch contains failing tests to be used as a discussion point for how we could/should modify the state machine / flags for DS.RecordArray / DS.ManyArray / DS.Model. As it stands, it appears that there are some inconsistencies (maybe intentional?) between them.

All of the tests which are passing have been modified to match their original passing conditions.

The new flagEquals() assertions and stateEquals() assertions should probably be audited to ensure that the state of the records should actually be what the assertion says they are. While re-writing these tests, I personally disagree with some of the flags set for the given record states.

Summary of changes:
- added comments for test helpers
- removed teardown for tests (all tests remained green after doing so)
- cleaned up scoping for tests -- globals only used for configuration, locals used for test data
- relocated helper functions to top of test
- removed unnecessary var definitions
- created a consistent test format:
- test setup code is grouped together
- test assertion code is grouped together
- whitespace between setup and assertion code
- comments above each individual assertion and setup block
- each test should have a state and flag assertion after a model change state