Add documentation for using underscored API endpoings and underscored root objects in JSON responses.
Starting in Beta 1 the RESTAdapter tries to send API requests to '/blog-posts' instead of '/blog_posts' as it had done previously.  It also expects the root object of a JSON response to be camelized instead of underscored.  This PR add info to the TRANSITION document about how to go back to the previous functionality.