Fix Travis failures and publishing of builds.
Updates to the latest ember-dev which ensures that the git config
user.name and git config user.email values are present before calling
git cherry-pick. This is due to a recent change in git 1.8.5 which
requires the user information be present prior to a cherry-pick.

Travis accidentally reverted to using an older git version that does not
support the `--points-at` flag that we are using in `ember-dev`.

This forces the system to upgrade to the latest version.

Travis ticket for reference:
https://github.com/travis-ci/travis-ci/issues/1710.