Remove underscores and rename
Having adapters and serializers with underscores in their names is
causing issues with Ember Appkit's Resolver. The resolver will normalize
the fullpath of the es6 module and calls `Ember.String.dashserize` on
the fullpath. This converts the underscores to dashes, for example:
`-ams` instead of `_ams` so the lookup becomes `app/serializers/-ams`
and this of course cannot be found.

This PR removes what can be argued as an unnecessary attempt to protect
the adapter and serializer names from being overwritten, instead the
convention to be pushed should be to have the names match the
adpter/serializer class name.

In addition, the ActiveModelAdapter's name was strangely called `_ams`
and this does't match the class name at all. There is no 's' in the
class name at all. Instead this PR has opted to rename both the adapter
and the serializer `active-model`.