Fix documentation for DS.Model.isNew
The act of finding a model in the store doest not implicitly save the record, thus the record.isNew property would still be true.  The record must be saved/committed to the server and if this promise is fulfilled, then it is no longer considered new.

If my assumption is incorrect, and store.find(...) is correct I would still prefer for a something else in code example since there is no clear indication of why finding a model with an id of 1 would imply to find the previously created record.