hasMany relationship property are essentially readOnly, lets mark them as such.
This also fixes a potential issue with minified code changing the arity of these CPs.
This happens because "value"'s only purpose is to increase the arity to 2, such that
setting to this CP has no effect. Unfortunately, when minified it does have an effect.

Also, allowing this property to be set to, even though it has no outcome, is just straight up misleading.