Refactor Ember initializer to use DS._setupContainer
The container setup used by Ember-Data has been coupled to the Ember.js
application initialization code. This splits the container setup into a
separate set of files, allowing container setup to be used in testing
environments that have not booted an application. Testing environments
should use the new `DS._setupContainer` API:

    var container = new Ember.Container();
    DS._setupContainer(container);
    var store = container.lookup('store:main');

`DS._setupContainer` will configure the container for ember-data proper
and for the activemode-adapter package.

This API uses an _ as a warning to consumers: It may change.