Add support for FormData in RESTAdapter#ajaxOptions.
This change allows users to pass an instance of FormData to any
adapter method the ends up making an ajax request. This is mostly
useful when creating or updating a record, since uploading files
becomes a much simpler matter.

Previously, ember-data would simply try to JSON.stringify
a FormData instance if it was passed, which is obviously
not the best behaviour.

A few unit tests have been added, but I didn't see any obvious
place to update the documentation. Any recommendations?