Favour declared mapping over keyForAttribute, if defined;
keyForAttribute, if defined, is useful for defining custom changes for all keys on a particular serializer, be it for the application or just a particular resource. The custom attr-based mapping should be maintained, and should precedence, over the keyForAttribute.

Broke down the ternary operators for increased readability.

Adds integration test for mapping > `keyForAttribute` precedence