PERF: O(n) -> O(1) record within recordArray check
If a record changes, we attempt to re-add it to the array, addRecord only adds unique records to the array but does so with an O(n) indexOf. A proper solution would be for addRecord to utilize an index, unfortunately it does not (yet). Luckily we already have just such an index, that records recordArrays map.