Do not clear own relationships when deleting records
If a record had a relationship, such as a `comment` having a `post`, previously
if you did `comment.deleteRecord()`, `comment.get('post')` would immediately be set
to `null`.

This behavior was wrong for several reasons:
  1.It prevented future rollback support for relationships
  2.If you had nested URLs, you would not be able to send a DELETE request to /post/1/comments/1
    because comment.get('post') would already have been deleted

We now only clear the inverse side of the relationships when the record is deleted.
The resulting refactor also enables us to add support for a record in a `404` state to be removed
from relationships and for removing records from manyArrays that do not have inverses.