fix addRecords error iterating over records
Reverts a change in commit c6b451c that used a
traditional for loop, assuming that “records” in
addRecords/removeRecords would always be a plain
JavaScript array. This assumption was inaccurate.

We use the Ember.EnumerableUtils to safely
iterate over the objects when forEach/etc aren’t
present. For example, IE8 with
Ember.EXTEND_PROTOTYPES = false;

fixes Github Issue #2533