Fixed the test "create - relationships are not duplicated". Since you will not predefine the id of a comment created on the client side, it does not make very much sense.
It seems like a good idea to add this failing test so it can be fixed. I am not very deep involved in ember data so I did not manage to solve it myself.

For example in ticket #1829 they describe the same problem. Also in this discussion: http://discuss.emberjs.com/t/two-ember-objects-created-for-single-backend-object/3891.