Implement ES6-compliant Date string parsing
As noted in the changeset, the ECMAScript 6 specification interprets
timezone-less Date strings according to the local system's time zone.
This is in direct opposition to ECMAScript 5 specification and the
currently-implemented behavior. From section 15.9.1.15 of the ES5
specification:

> the time zone offset specified as “Z” (for UTC) [...] The value of an
> absent time zone offset is “Z”.

Update Ember Data's Date string parsing logic to comply with the latest
specification.