Sort query params in ajax calls.
When developing APIs, is pretty common to place a high level caching
mechanism, like Varnish, to cache requests to public API endpoints, and
those tools use the URL string to determine if they have a cached
response.

As an example, if one user requests `/posts?sort=price&category=pets` and
another requests `/posts?category=pets&sort=price`, the cached request
won't be used in the second call.

Varnish by example has some modules to sort query parameters when
caching, but it's not the default behavior and requires custom config.

Although the URL specification does not say that the order of the query
parameters is meaningless, in the bast mayority (read: all) of the APIs
I've seen, 2 requests are considered identical regardless of the the order.

Because of that, sorting the query params consistenlty is considered a
good practice, so this PR introduces a `sortQueryParams` key in the RESTAdapter
class, that defaults to a function to sort query params alphabetically.

In case you don't want this behavior, setting that property to a falsey
value in the ApplicationAdapter will respect the original order.

In case you want a different order, you can pass a function that does
the sorting your way.

As other customization in ED, this can be customized on a per resource
level, in case the order matters in only a few endpoints.